# docs/protocol/openapi.yaml
...
components:
  schemas:
...
    CopyMoveRequest:
      properties:
        destination_path:
          default: ''
          title: Destination Path
          type: string
        destination_share_id:
          title: Destination Share Id
          type: string
        operation:
          default: copy
          enum:
          - copy
          - move
          title: Operation
          type: string
        source_path:
          default: ''
          title: Source Path
          type: string
        source_share_id:
          title: Source Share Id
          type: string
      required:
      - source_share_id
      - destination_share_id
      title: CopyMoveRequest
      type: object
    CreateFolderRequest:
      properties:
        name:
          maxLength: 255
          minLength: 1
          title: Name
          type: string
        path:
          default: ''
          title: Path
          type: string
      required:
      - name
      title: CreateFolderRequest
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    PairRequest:
      properties:
        client_id:
          maxLength: 128
          minLength: 8
          title: Client Id
          type: string
        client_name:
          maxLength: 80
          minLength: 1
          title: Client Name
          type: string
        pair_code:
          maxLength: 20
          minLength: 6
          pattern: ^[0-9]+$
          title: Pair Code
          type: string
      required:
      - client_id
      - client_name
      - pair_code
      title: PairRequest
      type: object
...
info:
  description: Internal transport between paired LanLink devices on a local network. This API never serves
    a user interface. Every deployment is HTTPS with a self-signed per-device certificate that peers pin
    at pairing time.
  title: LanLink v1
  version: '1.0'
openapi: 3.1.0
paths:
...
  /v1/clipboard:
    get:
      operationId: get_clipboard_v1_clipboard_get
      parameters:
      - in: header
        name: x-lanlink-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lanlink-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Clipboard V1 Clipboard Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Clipboard
    post:
      operationId: post_clipboard_v1_clipboard_post
      parameters:
      - in: header
        name: x-lanlink-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lanlink-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Post Clipboard V1 Clipboard Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Post Clipboard
...
  /v1/files/{share_id}:
    get:
      operationId: download_v1_files__share_id__get
      parameters:
      - in: path
        name: share_id
        required: true
        schema:
          title: Share Id
          type: string
      - in: query
        name: path
        required: true
        schema:
          title: Path
          type: string
      - in: header
        name: Range
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Range
      - in: header
        name: x-lanlink-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lanlink-Token
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Download
    put:
      description: 'Raw resumable upload; never buffers a whole file.


        Bytes land in a ``.lanlink-part`` sidecar. Only once the transfer

        completes (and matches its checksum, when one is supplied) does the file

        appear under its real name — so an interrupted transfer can never be

        mistaken for a finished one.'
      operationId: stream_upload_v1_files__share_id__put
      parameters:
      - in: path
        name: share_id
        required: true
        schema:
          title: Share Id
          type: string
      - in: query
        name: path
        required: false
        schema:
          default: ''
          title: Path
          type: string
      - in: query
        name: name
        required: true
        schema:
          title: Name
          type: string
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      - in: query
        name: finalize
        required: false
        schema:
          default: true
          title: Finalize
          type: boolean
      - in: query
        name: sha256
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Sha256
      - in: header
        name: x-lanlink-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lanlink-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Stream Upload V1 Files  Share Id  Put
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Stream Upload
  /v1/operations:
    post:
      operationId: operation_v1_operations_post
      parameters:
      - in: header
        name: x-lanlink-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lanlink-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CopyMoveRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Operation V1 Operations Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Operation
  /v1/pair:
    post:
      operationId: pair_v1_pair_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PairRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Pair V1 Pair Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Pair
  /v1/pairings/{client_id}:
    delete:
      operationId: revoke_pairing_v1_pairings__client_id__delete
      parameters:
      - in: path
        name: client_id
        required: true
        schema:
          title: Client Id
          type: string
      - in: header
        name: x-lanlink-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lanlink-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Revoke Pairing V1 Pairings  Client Id  Delete
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Revoke Pairing
  /v1/remote/media:
    post:
      operationId: remote_media_v1_remote_media_post
      parameters:
      - in: header
        name: x-lanlink-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lanlink-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Remote Media V1 Remote Media Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Remote Media
  /v1/remote/mouse:
    post:
      operationId: remote_mouse_v1_remote_mouse_post
      parameters:
      - in: header
        name: x-lanlink-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lanlink-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Remote Mouse V1 Remote Mouse Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Remote Mouse
  /v1/shares:
    get:
      operationId: shares_v1_shares_get
      parameters:
      - in: header
        name: x-lanlink-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lanlink-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Shares V1 Shares Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Shares
  /v1/shares/{share_id}/checksum:
    get:
      operationId: file_checksum_v1_shares__share_id__checksum_get
      parameters:
      - in: path
        name: share_id
        required: true
        schema:
          title: Share Id
          type: string
      - in: query
        name: path
        required: true
        schema:
          title: Path
          type: string
      - in: header
        name: x-lanlink-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lanlink-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response File Checksum V1 Shares  Share Id  Checksum Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: File Checksum
  /v1/shares/{share_id}/entries:
    delete:
      operationId: delete_v1_shares__share_id__entries_delete
      parameters:
      - in: path
        name: share_id
        required: true
        schema:
          title: Share Id
          type: string
      - in: query
        name: path
        required: true
        schema:
          title: Path
          type: string
      - in: query
        name: recursive
        required: false
        schema:
          default: false
          title: Recursive
          type: boolean
      - in: header
        name: x-lanlink-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lanlink-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Delete V1 Shares  Share Id  Entries Delete
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete
  /v1/shares/{share_id}/finalize:
    post:
      operationId: finalize_v1_shares__share_id__finalize_post
      parameters:
      - in: path
        name: share_id
        required: true
        schema:
          title: Share Id
          type: string
      - in: query
        name: name
        required: true
        schema:
          title: Name
          type: string
      - in: query
        name: path
        required: false
        schema:
          default: ''
          title: Path
          type: string
      - in: query
        name: sha256
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Sha256
      - in: header
        name: x-lanlink-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lanlink-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Finalize V1 Shares  Share Id  Finalize Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Finalize
  /v1/shares/{share_id}/folders:
    post:
      operationId: make_folder_v1_shares__share_id__folders_post
      parameters:
      - in: path
        name: share_id
        required: true
        schema:
          title: Share Id
          type: string
      - in: header
        name: x-lanlink-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lanlink-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFolderRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Make Folder V1 Shares  Share Id  Folders Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Make Folder
...
servers:
- url: https://{host}:{port}
  variables:
    host:
      default: 192.168.1.20
      description: LAN address from mDNS or an invite
    port:
      default: '8765'
      description: TCP port from mDNS or an invite